+++ MobSync/UIColor+Hex.m
@@ -12,7 +12,7 @@
 
 +(UIColor *)colorWithHex:(NSString *)hexString
 {
-    NSUInteger red, green, blue;
+    int red, green, blue;
     sscanf([hexString UTF8String], "#%02X%02X%02X", &red, &green, &blue);
     
     return [UIColor colorWithRed:red/255.0 green:green/255.0 blue:blue/255.0 alpha:1];
